backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dataclasses-json==0.6.1
Django==4.2.6
djangorestframework==3.14.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
filelock==3.13.0
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
huggingface-hub==0.17.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.324
langsmith==0.0.52
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
numpy==1.26.1
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.18.1
# nvidia-nvjitlink-cu12==12.3.52
# nvidia-nvtx-cu12==12.1.105
openai==0.28.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL==23.3.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.3
# sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
# transformers==4.34.1
triton==2.1.0
Twisted==23.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
zope.interface==6.1
